refactor: Dashboard JS split: extract tasks.js (per-domain split, round… [ORB-00149]
Dashboard JS split: extract tasks.js (per-domain split, round 1)

Planned-By: codex
