+++ crates/orbit-dashboard/src/lib.rs
@@ -27,6 +27,7 @@ use orbit_core::{OrbitError, OrbitRuntime};
 const INDEX_HTML: &str = include_str!("../assets/dashboard/index.html");
 const APP_JS: &str = include_str!("../assets/dashboard/app.js");
 const COMMON_JS: &str = include_str!("../assets/dashboard/common.js");
+const TASKS_JS: &str = include_str!("../assets/dashboard/tasks.js");
 
 /// Arguments for `orbit web serve` (and the library entry point).
 #[derive(Args, Clone)]
@@ -60,6 +61,7 @@ pub fn serve(runtime: &OrbitRuntime, args: ServeArgs) -> Result<(), OrbitError> 
         .route("/", get(serve_index))
         .route("/static/app.js", get(serve_app_js))
         .route("/static/common.js", get(serve_common_js))
+        .route("/static/tasks.js", get(serve_tasks_js))
         .route("/healthz", get(healthz))
         .nest("/api", api::router())
         .with_state(runtime);
@@ -121,6 +123,17 @@ async fn serve_common_js() -> Response {
             HeaderValue::from_static("application/javascript; charset=utf-8"),
         )],
         COMMON_JS,
+    )
+        .into_response()
+}
+
+async fn serve_tasks_js() -> Response {
+    (
+        [(
+            header::CONTENT_TYPE,
+            HeaderValue::from_static("application/javascript; charset=utf-8"),
+        )],
+        TASKS_JS,
     )
         .into_response()
 }
